--- test/test.ino
@@ -1,10 +1,0 @@
-void setup() {
-  // put your setup code here, to run once:
-  Serial.begin(9600);
-}
-
-void loop() {
-  // put your main code here, to run repeatedly:
-  int val = analogRead(A1);
-  Serial.println(val);
-}
